requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
ruff==0.0.286
soupsieve==2.4.1
urllib3==2.0.4
